+++ src/Cwaff-Guns/KingsLaw.cs
@@ -28,7 +28,8 @@ public class KingsLaw : CwaffGun
     public static void Init()
     {
         Lazy.SetupGun<KingsLaw>(ItemName, ShortDescription, LongDescription, Lore)
-          .SetAttributes(quality: ItemQuality.A, gunClass: GunClass.CHARGE, reloadTime: 0.75f, ammo: 1000, doesScreenShake: false, reloadAudio: "knife_gun_reload")
+          .SetAttributes(quality: ItemQuality.A, gunClass: GunClass.CHARGE, reloadTime: 0.75f, ammo: 1000, doesScreenShake: false,
+            reloadAudio: "knife_gun_reload", attacksThroughWalls: true)
           .AddToShop(ModdedShopType.TimeTrader)
           .InitProjectile(GunData.New(clipSize: 20, shootStyle: ShootStyle.Automatic, damage: 7.5f, speed: 40.0f, range: 999999f, force: 9f, customClip: true,
             cooldown: _SPAWN_RATE, sprite: "kings_law_projectile", fps: 12, scale: 0.5f, anchor: Anchor.MiddleCenter, spawnSound: "snd_undynedis", hitSound: "knife_gun_hit",