+++ src/Cwaff-Guns/Starmageddon.cs
@@ -16,7 +16,7 @@ public class Starmageddon : CwaffGun
     {
         Lazy.SetupGun<Starmageddon>(ItemName, ShortDescription, LongDescription, Lore)
           .SetAttributes(quality: ItemQuality.S, gunClass: GunClass.FULLAUTO, reloadTime: 1.0f, ammo: 900, shootFps: 20, chargeFps: 20,
-            reloadFps: 30, loopReloadAt: 0)
+            reloadFps: 30, loopReloadAt: 0, attacksThroughWalls: true)
           .InitProjectile(GunData.New(clipSize: 30, cooldown: 0.125f, angleVariance: 15.0f,
             shootStyle: ShootStyle.Automatic, damage: 8.0f, speed: 60.0f, range: 999999f, spawnSound: "starmageddon_fire_sound",
             sprite: "starmageddon_bullet", fps: 12, scale: 0.5f, anchor: Anchor.MiddleCenter, useDummyChargeModule: true, customClip: true,